Document script and Docker build context
Comments were to the script to make it more readable, especially to
those who wish to use it to futher development efforts.

Also, because I'm a pedant, I made sure to send the location of the
Docker file (hence, build context) along to the `docker build` command
so that it can be executed from anywhere.

Lastly, I set `xtrace` so that the `docker build` command is displayed
before it's executed; that way, one can audit the command when debugging
a build.