Add comments and a check for `jq`
If the `jq` command isn't located (via `which`) on the system, display
a helpful message to STDERR and exit with a failure.  Why?  Well, this
is a new requirement and I, personally, would rather see an explanation
why my previously working builds are now failing rather than either
having the build fail with an odd error message or, possibly worse, have
it "succeed" with garbage build arguments.

Also, I put a comment in at the top of the script explaining the purpose
of the script and why it's likely most users don't need to build the
image themselves.