Update README.md, to remove ">" from codeblock
Hey there! 
I love the project! It saves a lot of time, when starting new modules.

One "issue" I've had with it is the presense of a  ">" in the codeblock.
I catch myself copy-pasting it blindly more than once, so I decided to raise this PR, maybe it helps somebody in the future :)